Generalized the Register class and moved Register and Field classes to separate modules.
