+++ botblox_config/switch/ip175g.py
@@ -1,6 +1,9 @@
-from typing import List
+from typing import List, Optional
 
-from .switch import BitField, BitsField, ByteField, Port, PortListField, Register, ShortField, SwitchChip, SwitchFeature
+from .switch import SwitchChip, SwitchFeature
+from .port import Port
+from .register import MIIRegister, MIIRegisterAddress
+from .fields import BitField, BitsField, ByteField, PortListField, ShortField
 
 
 class IP175G(SwitchChip):
@@ -39,24 +42,24 @@ class IP175G(SwitchChip):
             self._ports.append(Port("5", 7))
 
     def _init_registers(self) -> None:
-        self._add_register(Register(23, 0, self))
-        self._add_register(Register(23, 1, self))
-        self._add_register(Register(23, 2, self))
+        self._add_register(MIIRegister(23, 0))
+        self._add_register(MIIRegister(23, 1))
+        self._add_register(MIIRegister(23, 2))
 
-        self._add_register(Register(23, 7, self))
-        self._add_register(Register(23, 8, self))
-        self._add_register(Register(23, 9, self))
+        self._add_register(MIIRegister(23, 7))
+        self._add_register(MIIRegister(23, 8))
+        self._add_register(MIIRegister(23, 9))
         if not self._nano:
-            self._add_register(Register(23, 11, self))
-            self._add_register(Register(23, 12, self))
+            self._add_register(MIIRegister(23, 11))
+            self._add_register(MIIRegister(23, 12))
 
-        self._add_register(Register(23, 13, self))
-        self._add_register(Register(23, 14, self))
+        self._add_register(MIIRegister(23, 13))
+        self._add_register(MIIRegister(23, 14))
 
-        self._add_register(Register(23, 19, self))
+        self._add_register(MIIRegister(23, 19))
 
         for i in range(25):
-            self._add_register(Register(24, i, self))
+            self._add_register(MIIRegister(24, i))
 
     def _init_fields(self) -> None:
         # Example values of the VLAN-related registries
@@ -73,44 +76,47 @@ class IP175G(SwitchChip):
         # [24, 2, 1, 0],  # VID_1=2
         # [24, 17, 0b10001000, 255],  # [VLAN_MEMBER_0=FF, VLAN_MEMBER_1=FF]
 
-        self._add_field(BitField(self._registers[23][0], 13, False, 'UNVID_MODE'))
-        self._add_field(BitField(self._registers[23][0], 15, False, 'VLAN_TABLE_CLR'))
+        Addr = MIIRegisterAddress
 
-        self._add_field(self._create_port_list_field(self._registers[23][1], 1, False, 'VLAN_CLS'))
-        self._add_field(self._create_port_list_field(self._registers[23][1], 0, False, "TAG_VLAN_EN"))
+        self._add_field(BitField(self._registers[Addr(23, 0)], 13, False, 'UNVID_MODE'))
+        self._add_field(BitField(self._registers[Addr(23, 0)], 15, False, 'VLAN_TABLE_CLR'))
 
-        self._add_field(BitField(self._registers[23][2], 13, False, 'VLAN_DROP_CFI'))
-        self._add_field(BitField(self._registers[23][2], 12, False, 'RSVD_VID_2'))
-        self._add_field(BitField(self._registers[23][2], 11, False, 'RSVD_VID_1'))
-        self._add_field(BitField(self._registers[23][2], 10, True, 'RSVD_VID_0'))
-        self._add_field(BitsField(self._registers[23][2], 8, 2, 0, 'ACCEPTABLE_FRM_TYPE'))
-        self._add_field(self._create_port_list_field(self._registers[23][2], 0, True, "VLAN_INGRESS_FILTER"))
+        self._add_field(self._create_port_list_field(self._registers[Addr(23, 1)], 1, False, 'VLAN_CLS'))
+        self._add_field(self._create_port_list_field(self._registers[Addr(23, 1)], 0, False, "TAG_VLAN_EN"))
 
-        self._add_field(ShortField(self._registers[23][7], 0, 1, 'VLAN_INFO_0'))
-        self._add_field(ShortField(self._registers[23][8], 0, 1, 'VLAN_INFO_1'))
-        self._add_field(ShortField(self._registers[23][9], 0, 1, 'VLAN_INFO_2'))
+        self._add_field(BitField(self._registers[Addr(23, 2)], 13, False, 'VLAN_DROP_CFI'))
+        self._add_field(BitField(self._registers[Addr(23, 2)], 12, False, 'RSVD_VID_2'))
+        self._add_field(BitField(self._registers[Addr(23, 2)], 11, False, 'RSVD_VID_1'))
+        self._add_field(BitField(self._registers[Addr(23, 2)], 10, True, 'RSVD_VID_0'))
+        self._add_field(BitsField(self._registers[Addr(23, 2)], 8, 2, 0, 'ACCEPTABLE_FRM_TYPE'))
+        self._add_field(self._create_port_list_field(self._registers[Addr(23, 2)], 0, True, "VLAN_INGRESS_FILTER"))
+
+        self._add_field(ShortField(self._registers[Addr(23, 7)], 0, 1, 'VLAN_INFO_0'))
+        self._add_field(ShortField(self._registers[Addr(23, 8)], 0, 1, 'VLAN_INFO_1'))
+        self._add_field(ShortField(self._registers[Addr(23, 9)], 0, 1, 'VLAN_INFO_2'))
         if not self._nano:
-            self._add_field(ShortField(self._registers[23][11], 0, 1, 'VLAN_INFO_3'))
-            self._add_field(ShortField(self._registers[23][12], 0, 1, 'VLAN_INFO_4'))
+            self._add_field(ShortField(self._registers[Addr(23, 11)], 0, 1, 'VLAN_INFO_3'))
+            self._add_field(ShortField(self._registers[Addr(23, 12)], 0, 1, 'VLAN_INFO_4'))
 
-        self._add_field(self._create_port_list_field(self._registers[23][13], 0, False, "ADD_TAG"))
-        self._add_field(self._create_port_list_field(self._registers[23][14], 0, False, "REMOVE_TAG"))
+        self._add_field(self._create_port_list_field(self._registers[Addr(23, 13)], 0, False, "ADD_TAG"))
+        self._add_field(self._create_port_list_field(self._registers[Addr(23, 14)], 0, False, "REMOVE_TAG"))
 
-        self._add_field(BitField(self._registers[23][19], 2, False, 'LEAKY_VLAN_2'))
-        self._add_field(BitField(self._registers[23][19], 1, False, 'LEAKY_VLAN_1'))
-        self._add_field(BitField(self._registers[23][19], 0, False, 'LEAKY_VLAN_0'))
+        self._add_field(BitField(self._registers[Addr(23, 19)], 2, False, 'LEAKY_VLAN_2'))
+        self._add_field(BitField(self._registers[Addr(23, 19)], 1, False, 'LEAKY_VLAN_1'))
+        self._add_field(BitField(self._registers[Addr(23, 19)], 0, False, 'LEAKY_VLAN_0'))
 
-        self._add_field(BitsField(self._registers[24][0], 0, 16, 0, 'VLAN_VALID'))
-
-        for i in range(16):
-            self._add_field(BitsField(self._registers[24][i + 1], 0, 12, i + 1, 'VID_{:1X}'.format(i)))
+        self._add_field(BitsField(self._registers[Addr(24, 0)], 0, 16, 0, 'VLAN_VALID'))
 
         for i in range(16):
-            self._add_field(self._create_port_list_field(self._registers[24][17 + (i // 2)], i % 2, True,
+            self._add_field(BitsField(self._registers[Addr(24, i + 1)], 0, 12, i + 1, 'VID_{:1X}'.format(i)))
+
+        for i in range(16):
+            self._add_field(self._create_port_list_field(self._registers[Addr(24, 17 + (i // 2))], i % 2, True,
                                                          "VLAN_MEMBER_{:1X}".format(i)))
 
     class IP175GPortListField(PortListField):
-        def __init__(self, register: Register, index: int, ports: List[Port], ports_default: bool, name: str) -> None:
+        def __init__(self, register: MIIRegister, index: int, ports: List[Port], ports_default: bool, name: str) \
+                -> None:
             super().__init__(register, ports, ports_default, name)
             self._base_field = ByteField(register, index, 255 if ports_default else 0, name)
             self._ports = ports if ports_default else list()
@@ -128,5 +134,26 @@ class IP175G(SwitchChip):
         def _clear(self, touch: bool = True) -> None:
             self._base_field.set_value(0, touch)
 
-    def _create_port_list_field(self, register: Register, index: int, ports_default: bool, name: str) -> PortListField:
+    def _create_port_list_field(self, register: MIIRegister, index: int, ports_default: bool, name: str) \
+            -> PortListField:
         return IP175G.IP175GPortListField(register, index, self._ports, ports_default, name)
+
+    def register_to_command(self, register: MIIRegister, leave_out_default: bool = True) -> Optional[List[int]]:
+        """
+        Convert the given register into a "command" for botblox firmware.
+        :param register: The register to convert.
+        :param leave_out_default: If true, returns None in case the register has its default value.
+        :return: The firmware "command".
+        """
+        if leave_out_default and register.is_default():
+            return None
+        return [register.address.phy, register.address.mii] + register.as_bytes()
+
+    def get_commands(self, leave_out_default: bool = True, only_touched: bool = False) -> List[List[int]]:
+        result = list()
+        for r in self._registers.values():
+            cmd = self.register_to_command(r, leave_out_default)
+            if cmd is not None and (not only_touched or r.is_touched()):
+                result.append(cmd)
+        result.sort()
+        return result
