+++ tests/switch/test_register.py
@@ -1,32 +1,11 @@
-from botblox_config.switch.switch import PortListField, Register, SwitchChip
-
-
-class ChipStub(SwitchChip):
-    def name(self) -> str:
-        return "stub"
-
-    def _init_features(self) -> None:
-        pass
-
-    def _init_ports(self) -> None:
-        pass
-
-    def _init_registers(self) -> None:
-        pass
-
-    def _init_fields(self) -> None:
-        pass
-
-    def _create_port_list_field(self, register: Register, index: int, ports_default: bool,
-                                name: str) -> PortListField:
-        pass
+from botblox_config.switch.register import MIIRegister
 
 
 class TestRegister:
     def test_register_zero(self) -> None:
-        r = Register(1, 2, ChipStub())
-        assert r.phy == 1
-        assert r.mii == 2
+        r = MIIRegister(1, 2)
+        assert r.address.phy == 1
+        assert r.address.mii == 2
         assert len(r.as_bytes()) == 2
         assert r.as_bytes()[0] == 0
         assert r.as_bytes()[1] == 0
@@ -64,10 +43,10 @@ class TestRegister:
             pass
 
     def test_register_set_bit_in_first_byte(self) -> None:
-        r = Register(3, 4, ChipStub())
+        r = MIIRegister(3, 4)
         r.set_bit(0, True)
-        assert r.phy == 3
-        assert r.mii == 4
+        assert r.address.phy == 3
+        assert r.address.mii == 4
         assert len(r.as_bytes()) == 2
         assert r.as_bytes()[0] == 1
         assert r.as_bytes()[1] == 0
@@ -83,7 +62,7 @@ class TestRegister:
         assert r.is_default()
 
     def test_register_set_bit_in_second_byte(self) -> None:
-        r = Register(3, 4, ChipStub())
+        r = MIIRegister(3, 4)
         r.set_bit(10, True)
         assert r.as_bytes()[0] == 0
         assert r.as_bytes()[1] == 0b0000_0100
@@ -100,7 +79,7 @@ class TestRegister:
         assert r.as_number(signed=True) == 0b0000_0100_0000_0000
 
     def test_register_set_two_bits(self) -> None:
-        r = Register(3, 4, ChipStub())
+        r = MIIRegister(3, 4)
         r.set_bit(2, True)
         r.set_bit(10, True)
         assert r.as_bytes()[0] == 0b0000_0100
@@ -120,7 +99,7 @@ class TestRegister:
         assert r.as_number(signed=True) == 0b0000_0100_0000_0100
 
     def test_register_reset_bit(self) -> None:
-        r = Register(3, 4, ChipStub())
+        r = MIIRegister(3, 4)
         r.set_bit(2, True)
         r.set_bit(2, False)
         assert r.as_bytes()[0] == 0
@@ -137,13 +116,13 @@ class TestRegister:
         assert r.as_number(signed=True) == 0
 
     def test_register_set_bit_as_signed(self) -> None:
-        r = Register(3, 4, ChipStub())
+        r = MIIRegister(3, 4)
         r.set_bit(15, True)
         assert r.as_number() == pow(2, 15)
         assert r.as_number(signed=True) == -pow(2, 15)
 
     def test_register_set_first_byte(self) -> None:
-        r = Register(3, 4, ChipStub())
+        r = MIIRegister(3, 4)
         r.set_byte(0, 100)
         assert len(r.as_bytes()) == 2
         assert r.as_bytes()[0] == 100
@@ -155,7 +134,7 @@ class TestRegister:
         assert r.as_number(signed=True) == 100
 
     def test_register_set_second_byte(self) -> None:
-        r = Register(3, 4, ChipStub())
+        r = MIIRegister(3, 4)
         r.set_byte(1, 0xCC)
         assert len(r.as_bytes()) == 2
         assert r.as_bytes()[0] == 0
@@ -167,7 +146,7 @@ class TestRegister:
         assert r.as_number(signed=True) == -13312
 
     def test_register_set_both_bytes(self) -> None:
-        r = Register(3, 4, ChipStub())
+        r = MIIRegister(3, 4)
         r.set_byte(0, 0xCC)
         r.set_byte(1, 0xDD)
         assert len(r.as_bytes()) == 2
@@ -179,7 +158,7 @@ class TestRegister:
         assert r.as_number() == 0xDDCC
 
     def test_register_set_bits(self) -> None:
-        r = Register(3, 4, ChipStub())
+        r = MIIRegister(3, 4)
         r.set_bits(6, 4, 0b1010)
         assert r.as_bytes()[0] == 0b1000_0000
         assert r.as_bytes()[1] == 0b0000_0010
@@ -195,7 +174,7 @@ class TestRegister:
         assert r.as_number() == 0b0000_0010_1000_0000
 
     def test_register_set_all_bits(self) -> None:
-        r = Register(3, 4, ChipStub())
+        r = MIIRegister(3, 4)
         r.set_bits(0, 16, 0b1010_0101_1100_1111)
         assert r.as_bytes()[0] == 0b1100_1111
         assert r.as_bytes()[1] == 0b1010_0101
